[DITP-pilotage/pilote-2] feat(PIL-1370): brancher le feature flag REORGANISATION_PAGE_ACCUEIL
Crée PageAccueilLegacy (copie exacte de l'ancien ChantierLayout inline) et modifie
index.tsx pour switcher entre PageAccueil (nouvelle architecture en sections) et
PageAccueilLegacy selon le feature flag NEXT_PUBLIC_FF_REORGANISATION_PAGE_ACCUEIL.
Les constantes PROFIL_* sont déplacées dans BasePageAccueilLayout et PageAccueilLegacy.

Refs: PIL-1370
Project: pilote-2
Client: DITP-pilotage
Category: feature
Stack: react, typescript
Tags: frontend, feature-flag
Files-changed: 2
