+++ src/pages/accueil/chantier/[territoireCode]/index.tsx
@@ -1,42 +1,21 @@
-import { useState } from "react";
 import { GetServerSidePropsContext, InferGetServerSidePropsType } from "next";
-import Head from "next/head";
-import { useSession } from "next-auth/react";
-import { parseAsStringLiteral, useQueryState } from "nuqs";
 import assert from "node:assert/strict";
-import { useProfilUtilisateurConnecte } from "@/client/hooks/useProfilUtilisateurConnecte";
 import { useEnv } from "@/client/hooks/useEnv";
-import PageChantiers from "@/components/PageAccueil/PageChantiers/PageChantiers";
-import BarreLatérale from "@/components/_commons/BarreLatérale/BarreLatérale";
-import BarreLatéraleEncart from "@/components/_commons/BarreLatérale/BarreLatéraleEncart/BarreLatéraleEncart";
-import { Filtres } from "@/components/PageAccueil/Filtres/Filtres";
 import { auth } from "@/server/infrastructure/api/auth/[...nextauth]";
 import Axe from "@/server/domain/axe/Axe.interface";
 import Alerte from "@/server/domain/alerte/Alerte";
 import { presenterEnAvancementsStatistiquesAccueilContrat } from "@/server/chantiers/app/contrats/AvancementsStatistiquesAccueilContrat";
 import { objectEntries } from "@/client/utils/objects/objects";
-import { ProfilEnum } from "@/server/app/enum/profil.enum";
 import { territoireCodeVersMailleCodeInsee } from "@/server/utils/territoires";
 import { Chantier } from "@/server/chantiers/domain/Chantier";
 import { FiltreQueryParams } from "@/server/chantiers/app/contrats/FiltreQueryParams";
 import { presenterEnRépartitionsMétéosChantiersContrat } from "@/server/chantiers/app/contrats/RepartitionMeteoChantiersContrat";
 import { getAnneeDateDeBascule } from "@/components/_commons/IndicateursChantier/Bloc/ValeurEtDate/getAnneeDateDeBascule";
 import { configuration, configurationFeatureFlip } from "@/config";
-import { ModaleVideoAccueil } from "@/components/PageAccueil/PageChantiers/ModaleVideoAccueil/ModaleVideoAccueil";
 import { getContainer } from "@/server/dependances";
-import { profilsRégionaux } from "@/server/gestion-utilisateur/domain/Utilisateur.interface";
-import Titre from "@/components/_commons/Titre/Titre";
-import { estAutoriséAConsulterLaFicheTerritoriale } from "@/client/utils/fiche-territoriale/fiche-territoriale";
-import { PanelMenuNavigation } from "@/components/_commons/PanelMenuNavigation/PanelMenuNavigation";
-import { FiltresActifs } from "@/components/PageAccueil/FiltresActifs/FiltresActifs";
-import { ModaleInscriptionInfolettre } from "@/components/PageAccueil/PageChantiers/ModaleInscriptionInfoLettre/ModaleInscriptionInfolettre";
-import { ModaleRenseignerService } from "@/components/PageAccueil/PageChantiers/ModaleRenseignerService/ModaleRenseignerService";
-import { BoutonNavigationFicheTerritoriale } from "@/components/PageAccueil/BoutonNavigationFicheTerritoriale";
-import { BoutonNavigationRapportDetaille } from "@/components/BoutonNavigationRapportDetaille";
-import { BoutonSyntheseTerritoire } from "@/components/PageAccueil/BoutonSyntheseTerritoire";
-import { BoutonExportDesDonnees } from "@/components/PageAccueil/BoutonExportDesDonnees";
-import { clsxm } from "@/utils/clsxm";
 import { loadAccueilSearchParams } from "@/client/searchParams/accueilSearchParams";
+import { PageAccueil } from "@/components/PageAccueil/PageAccueil";
+import { PageAccueilLegacy } from "@/components/PageAccueil/PageAccueilLegacy";
 
 export const getServerSideProps = async (
   context: GetServerSidePropsContext,
@@ -297,204 +276,17 @@ export const getServerSideProps = async (
   };
 };
 
-const PROFIL_AUTORISE_A_VOIR_FILTRE_TERRITORIALISE = new Set([
-  ProfilEnum.CABINET_MTFP,
-  ProfilEnum.PM_ET_CABINET,
-  ProfilEnum.PR,
-  ProfilEnum.CABINET_MINISTERIEL,
-  ProfilEnum.DIR_ADMIN_CENTRALE,
-  ProfilEnum.DROM,
-  ProfilEnum.SECRETARIAT_GENERAL,
-  ProfilEnum.DIR_PROJET,
-  ProfilEnum.EQUIPE_DIR_PROJET,
-  ProfilEnum.DITP_ADMIN,
-  ProfilEnum.DITP_PILOTAGE,
-  ProfilEnum.PREFET_REGION,
-  ProfilEnum.COORDINATEUR_REGION,
-  ProfilEnum.SERVICES_DECONCENTRES_REGION,
-  ProfilEnum.RESPONSABLE_REGION,
-]);
-
-const PROFIL_REGIONAUX_AUTORISE_A_VOIR_FILTRE_TERRITORIALISE = new Set(
-  profilsRégionaux,
-);
-
-const ChantierLayout = ({
-  chantiers,
-  chantierIds,
-  chantierIdsSansFiltrageAlertes,
-  nombreTotalChantiersAvecAlertes,
-  axes,
-  ministères,
-  territoireCode,
-  mailleSelectionnee,
-  mailleQuery,
-  filtresComptesCalculés,
-  avancementsAgrégés,
-  avancementsGlobauxTerritoriauxMoyens,
-  repartitionMeteosChantiers,
-  jalon,
-  jalonParDefaut,
-  doitAfficherModaleVideoAccueil,
-  doitAfficherLaModaleInfolettre,
-  doitAfficherLaFicheTerritoriale,
-  moyenneTerritoire,
-}: InferGetServerSidePropsType<typeof getServerSideProps>) => {
-  const { data: session } = useSession();
-
-  const estProfilTerritorialise =
-    PROFIL_AUTORISE_A_VOIR_FILTRE_TERRITORIALISE.has(session?.profil || "");
-  const estProfilRegionalAutoriseAVoirLaTerritorialisation =
-    PROFIL_REGIONAUX_AUTORISE_A_VOIR_FILTRE_TERRITORIALISE.has(
-      session?.profil || "",
-    );
-
-  const profil = useProfilUtilisateurConnecte();
-  const monProfilEstDisponible = useEnv("NEXT_PUBLIC_FF_MON_PROFIL");
-  const ffAskAI = useEnv("NEXT_PUBLIC_FF_ASK_AI");
-  const doitAfficherModaleRenseignerService =
-    !!monProfilEstDisponible &&
-    (profil.service == null || profil.fonction == null);
-
-  const [estOuverteBarreLatérale, setEstOuverteBarreLatérale] = useState(false);
-  const [isModaleInfolettreOpen, setIsModaleInfolettreOpen] = useState(
-    doitAfficherLaModaleInfolettre,
+const ChantierLayout = (
+  props: InferGetServerSidePropsType<typeof getServerSideProps>,
+) => {
+  const ffReorganisationPageAccueil = useEnv(
+    "NEXT_PUBLIC_FF_REORGANISATION_PAGE_ACCUEIL",
   );
-  const [isModaleVideoAccueilOpen, setIsModaleVideoAccueilOpen] = useState(
-    doitAfficherModaleVideoAccueil,
-  );
-  const [isModaleRenseignerServiceOpen, setIsModaleRenseignerServiceOpen] =
-    useState(doitAfficherModaleRenseignerService);
 
-  const filtresStatut = useQueryState(
-    "statut",
-    parseAsStringLiteral([
-      "BROUILLON",
-      "PUBLIE",
-      "BROUILLON_ET_PUBLIE",
-      "ARCHIVE",
-    ]),
-  );
-  const chantiersSontArchives = filtresStatut?.includes("ARCHIVE") ?? false;
-  const PROFIL_INTERDIT_DE_VOIR_LE_SELECTEUR_DE_MAILLE = new Set([
-    ProfilEnum.COORDINATEUR_DEPARTEMENT,
-    ProfilEnum.RESPONSABLE_DEPARTEMENT,
-  ]);
-  const estAutoriseAVoirLeSelecteurDeMaille =
-    !PROFIL_INTERDIT_DE_VOIR_LE_SELECTEUR_DE_MAILLE.has(session?.profil || "");
-
-  const pathname = "/accueil/chantier/[territoireCode]";
-
-  return (
-    <div className="[&_.fr-h2]:!text-[1.875rem] [&_.fr-h2]:!leading-9">
-      <Head>
-        <title>PILOTE - Piloter l'action publique par les résultats</title>
-      </Head>
-      <div className="flex">
-        <BarreLatérale
-          estOuvert={estOuverteBarreLatérale}
-          setEstOuvert={setEstOuverteBarreLatérale}
-        >
-          <BarreLatéraleEncart>
-            <Titre
-              baliseHtml="h1"
-              className={clsxm(
-                `fr-h2 fr-p-0 fr-mb-3w`,
-                chantiersSontArchives
-                  ? "titre-gris"
-                  : "fr-text-title--blue-france",
-              )}
-            >
-              {`${nombreTotalChantiersAvecAlertes} ${nombreTotalChantiersAvecAlertes >= 2 ? "chantiers" : "chantier"}`}
-            </Titre>
-            <div className="inline-flex flex-column gap-1">
-              {doitAfficherLaFicheTerritoriale &&
-              estAutoriséAConsulterLaFicheTerritoriale(
-                session?.profil || "",
-              ) ? (
-                <BoutonNavigationFicheTerritoriale
-                  jalon={jalon}
-                  territoireCode={territoireCode}
-                />
-              ) : null}
-              <BoutonNavigationRapportDetaille
-                territoireCode={territoireCode}
-              />
-              <BoutonExportDesDonnees territoireCode={territoireCode} />
-            </div>
-          </BarreLatéraleEncart>
-          <section>
-            <Filtres
-              afficherToutLesFiltres
-              axes={axes}
-              estProfilRegionalAutoriseAVoirLaTerritorialisation={
-                estProfilRegionalAutoriseAVoirLaTerritorialisation
-              }
-              estProfilTerritorialise={estProfilTerritorialise}
-              ministères={ministères}
-            />
-          </section>
-        </BarreLatérale>
-        <div className="w-full">
-          <div className="sticky top-0 z-[1] w-full shadow-[0_6px_18px_var(--shadow-color)] bg-dsfr-blue-france-850 fr-grid-row fr-pt-2w">
-            <PanelMenuNavigation
-              estAutoriseAVoirLeSelecteurDeMaille={
-                estAutoriseAVoirLeSelecteurDeMaille
-              }
-              mailleQuery={mailleQuery}
-              pathname={pathname}
-              setEstOuverteBarreLatérale={setEstOuverteBarreLatérale}
-              territoireCode={territoireCode}
-            />
-            <FiltresActifs
-              axes={axes}
-              mailleSelectionnee={mailleSelectionnee}
-              ministères={ministères}
-            />
-            {ffAskAI || session?.profil === ProfilEnum.DITP_ADMIN ? (
-              <BoutonSyntheseTerritoire
-                territoireCode={territoireCode}
-                jalon={jalon}
-              />
-            ) : null}
-          </div>
-          <PageChantiers
-            avancementsAgrégés={avancementsAgrégés}
-            avancementsGlobauxTerritoriauxMoyens={
-              avancementsGlobauxTerritoriauxMoyens
-            }
-            chantierIds={chantierIds}
-            chantierIdsSansFiltrageAlertes={chantierIdsSansFiltrageAlertes}
-            chantiers={chantiers}
-            filtresComptesCalculés={filtresComptesCalculés}
-            jalon={jalon}
-            jalonParDefaut={jalonParDefaut}
-            mailleQuery={mailleQuery}
-            ministères={ministères}
-            nombreTotalChantiersAvecAlertes={nombreTotalChantiersAvecAlertes}
-            repartitionMeteosChantiers={repartitionMeteosChantiers}
-            territoireCode={territoireCode}
-            moyenneTerritoire={moyenneTerritoire}
-          />
-          <ModaleVideoAccueil
-            onOpenChange={setIsModaleVideoAccueilOpen}
-            open={isModaleVideoAccueilOpen}
-          />
-          <ModaleRenseignerService
-            onOpenChange={setIsModaleRenseignerServiceOpen}
-            open={!isModaleVideoAccueilOpen && isModaleRenseignerServiceOpen}
-          />
-          <ModaleInscriptionInfolettre
-            onOpenChange={setIsModaleInfolettreOpen}
-            open={
-              !isModaleVideoAccueilOpen &&
-              !isModaleRenseignerServiceOpen &&
-              isModaleInfolettreOpen
-            }
-          />
-        </div>
-      </div>
-    </div>
+  return ffReorganisationPageAccueil ? (
+    <PageAccueil {...props} />
+  ) : (
+    <PageAccueilLegacy {...props} />
   );
 };
